+++ Gopkg.lock
@@ -1,29 +1,9 @@
 # This file is autogenerated, do not edit; changes may be undone by the next 'dep ensure'.
 
 
-[[projects]]
-  branch = "master"
-  digest = "1:d74b398854b523dd3ae7fa3deb67422bfbd6e3beaf296fea9338a8722561d976"
-  name = "github.com/itsubaki/q"
-  packages = [
-    "internal/circuit/gate",
-    "internal/circuit/qubit",
-    "internal/math/matrix",
-    "internal/math/number",
-    "internal/math/vector",
-  ]
-  pruneopts = "UT"
-  revision = "63e8b222b2a2485767a9d3052209f03439422015"
-
 [solve-meta]
   analyzer-name = "dep"
   analyzer-version = 1
-  input-imports = [
-    "github.com/itsubaki/q/internal/circuit/gate",
-    "github.com/itsubaki/q/internal/circuit/qubit",
-    "github.com/itsubaki/q/internal/math/matrix",
-    "github.com/itsubaki/q/internal/math/number",
-    "github.com/itsubaki/q/internal/math/vector",
-  ]
+  input-imports = []
   solver-name = "gps-cdcl"
   solver-version = 1
