+++ gate/gate_test.go
@@ -5,7 +5,7 @@ import (
 	"math/cmplx"
 	"testing"
 
-	"github.com/itsubaki/q/internal/math/matrix"
+	"github.com/axamon/q/matrix"
 )
 
 func TestInverseU(t *testing.T) {
