+++ number/gcd.go
