+++ q/q.go
@@ -3,9 +3,9 @@ package q
 import (
 	"math"
 
-	"github.com/itsubaki/q/internal/circuit/gate"
-	"github.com/itsubaki/q/internal/circuit/qubit"
-	"github.com/itsubaki/q/internal/math/matrix"
+	"github.com/axamon/q/gate"
+	"github.com/axamon/q/matrix"
+	"github.com/axamon/q/qubit"
 )
 
 type Q struct {
