+++ qubit/probability.go
