+++ qubit/qubit_test.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/itsubaki/q/internal/circuit/gate"
-	"github.com/itsubaki/q/internal/math/matrix"
+	"github.com/axamon/q/gate"
+	"github.com/axamon/q/matrix"
 )
 
 func TestOuterProduct(t *testing.T) {
